+++ changelog.md
@@ -23,6 +23,7 @@
 1. add latest clients in dashboard
 2. implement sorting and pagination of clients
 3. pre-fill client selector when creating an invoice from client profile
+4. add print icon to invoice table options
 
 # 2022 03 22
 - invalidate/modify pages when invoice is created.